refactor: 피드백 step2 반영
