+++ README.md
@@ -12,3 +12,7 @@
 - [x] 로또 숫자는 작은 수부터 큰 수 순서로 정렬된다.
 - [x] 당첨은 3개의 숫자를 맞춘 사람부터 6개 맞춘 사람까지이다.
 - [x] 수익률은 로또 구매 금액 대비 당첨 금액이다.
+
+3. 로또(2등) 기능
+- [x] 보너스 볼을 입력한다.
+- [ ] 5개 일치하고 보너스 볼까지 일치하는 등급이 있다.
