+++ src/main/kotlin/lotto/domain/Winner.kt
@@ -1,10 +1,11 @@
-package lotto.enums
+package lotto.domain
 
 enum class Winner(
     val reward: Int,
     val matchCount: Int
 ) {
     FIRST_PLACE(2_000_000_000, 6),
+    SECOND_PLACE_WITH_BONUS(30_000_000, 6),
     SECOND_PLACE(1_500_000, 5),
     THIRD_PLACE(50_000, 4),
     FOURTH_PLACE(5_000, 3);
