+++ src/main/kotlin/lotto/view/InputView.kt
@@ -11,4 +11,9 @@ class InputView {
         println("지난 주 당첨 번호를 입력해주세요.")
         return readln().split(",").map { it.toInt() }
     }
+
+    fun inputBonusNumber(): Int {
+        println("보너스 볼을 입력해주세요.")
+        return readln().toInt()
+    }
 }
